requirements.txt
arrow
Pyarrow
PyPDF2
PyMuPDF
python-dotenv
tenacity
tiktoken
unstructured[all-docs]
xata